moved expand_options in util
